backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pandas==3.0.1
openpyxl==3.1.5
SQLAlchemy==2.0.48
psycopg[binary]==3.3.3
python-multipart==0.0.22
pydantic-settings==2.8.1
